Switch to app router
