+++ next.config.js
@@ -1,8 +1,6 @@
 /** @type {import('next').NextConfig} */
-const { i18n } = require('./next-i18next.config');
 
 module.exports = {
-  i18n,
   reactStrictMode: true,
   compiler: {
     styledComponents: true,
