+++ package.json
@@ -3,30 +3,31 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "next": "^13.1.6",
-    "next-i18next": "^12.0.1",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.5.0",
-    "react-dom": "^18.2.0",
-    "styled-components": "^5.3.5"
+    "@fortawesome/fontawesome-svg-core": "1.2.36",
+    "@fortawesome/free-brands-svg-icons": "5.15.4",
+    "@fortawesome/free-solid-svg-icons": "5.15.4",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "bootstrap": "5.3.2",
+    "next": "latest",
+    "react": "18.2.0",
+    "react-bootstrap": "2.5.0",
+    "react-dom": "18.2.0",
+    "styled-components": "6.0.8"
   },
   "devDependencies": {
-    "@types/jest": "^29.0.2",
-    "@types/node": "^18.7.18",
-    "@types/react": "^18.0.20",
-    "@types/react-bootstrap": "^0.32.30",
-    "@types/react-dom": "^18.0.6",
-    "@types/styled-components": "^5.1.26",
-    "@typescript-eslint/eslint-plugin": "^5.50.0",
-    "eslint": "^8.33.0",
-    "eslint-config-next": "^13.1.6",
-    "eslint-config-prettier": "^8.6.0",
-    "prettier": "^2.7.1",
-    "typescript": "^4.8.3"
+    "@types/jest": "29.5.5",
+    "@types/node": "20.7.2",
+    "@types/react": "18.2.23",
+    "@types/react-bootstrap": "0.32.33",
+    "@types/react-dom": "18.2.8",
+    "@types/styled-components": "5.1.28",
+    "@typescript-eslint/eslint-plugin": "6.7.3",
+    "@typescript-eslint/parser": "6.7.3",
+    "eslint": "8.50.0",
+    "eslint-config-next": "13.5.3",
+    "eslint-config-prettier": "9.0.0",
+    "prettier": "3.0.3",
+    "typescript": "5.2.2"
   },
   "scripts": {
     "dev": "next dev",
