--- pages/_document.tsx
@@ -1,32 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document';
-import i18nextConfig from '../next-i18next.config';
-
-const Document = () => {
-  const currentLocale = i18nextConfig.i18n.defaultLocale;
-
-  return (
-    <Html lang={currentLocale}>
-      <Head>
-        <link rel="shortcut icon" href="/favicon.ico" />
-        <link rel="manifest" href="/manifest.json" />
-        <link
-          rel="stylesheet"
-          href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-          integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-          crossOrigin="anonymous"
-        />
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link
-          href="https://fonts.googleapis.com/css?family=Josefin+Slab&display=swap"
-          rel="stylesheet"
-        />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-};
-
-export default Document;
